requirements.txt
mysql-connector-python
pandas
openai
langchain-openai
llama-index
pickle
zlib